refactor: update LogFile import paths and create log-file module
